Refactor holes panel webview
