--- src/holesPanel/webview/holeState.ts
@@ -1,4 +1,0 @@
-export interface HoleState {
-  expanded: boolean;
-  highlighted: boolean;
-}
